fix: 모달 폴더 수정
