+++ client/src/components/banner/MainBanner.tsx
@@ -6,7 +6,7 @@ import { PiMagicWandFill } from 'react-icons/pi';
 import { PiCaretRightThin } from 'react-icons/pi';
 import { useNavigate } from 'react-router-dom';
 import { useModal } from '../../hooks/useModal';
-import Recommend from '../modal/Recommend';
+import Recommend from '../ui/questions/Recommend';
 
 function MainBanner() {
   const navigate = useNavigate();
